Refactor code structure and split Linux specific implementation
Break down the main.c file in two other files common.c and linux.c.
The common.c file contains functions which are common for all
POSIX-based targets and linux.c contains the implementations which are
Linux specific.

The goal of this refactor is to make urunit code easier to adapt for
other tare=get OSes like BSDs.
